Rename Services to Stores in Dapper module
