+++ src/modules/Elsa.Dapper/Modules/Runtime/Extensions/DapperWorkflowManagementFeatureExtensions.cs
@@ -1,5 +1,5 @@
 ﻿using Elsa.Dapper.Modules.Runtime.Features;
-using Elsa.Dapper.Modules.Runtime.Services;
+using Elsa.Dapper.Modules.Runtime.Stores;
 using Elsa.Workflows.Runtime.Features;
 
 // ReSharper disable once CheckNamespace
