+++ src/modules/Elsa.Dapper/Modules/Runtime/Features/DapperExecutionLogPersistenceFeature.cs
@@ -1,5 +1,5 @@
 using Elsa.Dapper.Features;
-using Elsa.Dapper.Modules.Runtime.Services;
+using Elsa.Dapper.Modules.Runtime.Stores;
 using Elsa.Features.Abstractions;
 using Elsa.Features.Attributes;
 using Elsa.Features.Services;
